modify project to reflect rails conventions; enable and fix testing environment; upgrade to rails 5
