+++ app/views/info_pages/_about_section.html.erb
