+++ app/views/info_pages/_intro_section.html.erb
