+++ app/views/info_pages/_projects_section.html.erb
