+++ app/views/info_pages/_team_section.html.erb
