+++ app/views/info_pages/index.html.erb
