+++ app/views/layouts/application.html.erb
@@ -19,6 +19,12 @@
   	</div>
   	<!-- Yield content form invidividual pages -->
 	<%= yield %>
+
+	<!-- Enable debug environment if in development environment --> 
+	<div class="container">
+    	<%= debug(params) if Rails.env.development? %>
+ 	</div>
+ 
 	<!-- Global Footer -->
 	<%= render "layouts/footer" %>
 </body>
