+++ app/views/members/new.html.erb
@@ -1,7 +1,7 @@
 <h1>New Member Page</h1>
 <p>Find me in app/views/member/new.html.erb</p>
 
-<%= form_for :member, url: member_index_path do |f| %>
+<%= form_for@member do |f| %>
     <p>
         <%= f.label :firstName %><br>
         <%= f.text_field :firstName %><br>
