+++ app/views/members/show.html.erb
