+++ app/views/projects/delete.html.erb
