+++ test/controllers/info_pages_controller_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class HomeControllerTest < ActionController::TestCase
+class InfoPagesControllerTest < ActionController::TestCase
   test "should get index" do
     get :index
     assert_response :success
