--- test/controllers/member_controller_test.rb
@@ -1,29 +1,0 @@
-require 'test_helper'
-
-class MemberControllerTest < ActionController::TestCase
-  test "should get new" do
-    get :new
-    assert_response :success
-  end
-
-  test "should get create" do
-    get :create
-    assert_response :success
-  end
-
-  test "should get show" do
-    get :show
-    assert_response :success
-  end
-
-  test "should get edit" do
-    get :edit
-    assert_response :success
-  end
-
-  test "should get delete" do
-    get :delete
-    assert_response :success
-  end
-
-end
