--- test/controllers/project_controller_test.rb
@@ -1,19 +1,0 @@
-require 'test_helper'
-
-class ProjectControllerTest < ActionController::TestCase
-  test "should get create" do
-    get :create
-    assert_response :success
-  end
-
-  test "should get edit" do
-    get :edit
-    assert_response :success
-  end
-
-  test "should get delete" do
-    get :delete
-    assert_response :success
-  end
-
-end
