+++ test/controllers/projects_controller_test.rb
@@ -1,0 +1,9 @@
+require 'test_helper'
+
+class ProjectsControllerTest < ActionController::TestCase
+  
+  test "should get new" do
+    get :new
+    assert_response :success
+  end
+end
